requirements.txt
openpyxl
requests
pandas
Pillow
pyinstaller